minor cleanups
